requirements.txt
# Utilidades gerais

Django>=1.8
django-debug-toolbar>=1.3.0
django-extensions>=1.5.2
ipdb>=0.8
ipython>=3.1.0
psycopg2>=2.6
Werkzeug>=0.10.4

# Específicos da aplicação

markdown
django-filter
djangorestframework
django-rest-auth

# Para o allauth funcionar no Django 1.8, só pegando a versão mais nova do github

#https://github.com/pennersr/django-allauth/archive/master.zip